+++ modules/System/assets/vue-components/fields/renderer.js
@@ -577,7 +577,7 @@ export default {
                             <icon class="kiss-size-5 kiss-color-muted kiss-margin-xsmall-left" v-if="field.i18n && locales.length" :title="t('Localized')">language</icon>
                             <icon class="kiss-size-5 kiss-color-danger kiss-margin-xsmall-left" v-if="field.required" :title="t('Required')">trip_origin</icon>
                         </label>
-                        <div class="kiss-flex kiss-flex-middle app-fieldcontainer-visible-hover" gap="small" v-if="fieldActions.length && !field.i18n">
+                        <div class="kiss-flex kiss-flex-middle app-fieldcontainer-visible-hover" gap="xsmall" v-if="fieldActions.length && !field.i18n">
                             <component :is="action.component" v-model="val[field.name]" :document="val" :field="field" :locale="locale" v-for="action in fieldActions"></component>
                         </div>
                         <a class="app-fieldcontainer-visible-hover" :class="{'kiss-color-muted': nested}" @click="clear(field, val)" :aria-label="t('Clear') + ': ' + (field.label || field.name)" kiss-tooltip="right" v-if="field.opts && !field.opts.readonly"><icon>backspace</icon></a>
@@ -592,9 +592,9 @@ export default {
                 <div class="kiss-margin-small-top" v-if="field.i18n && locales.length">
                     <div class="kiss-margin" v-for="locale in visibleLocales">
                         <div class="kiss-flex kiss-flex-middle" gap="small">
-                            <div class="kiss-margin-small kiss-flex kiss-flex-middle kiss-visible-toggle" v-if="Array.isArray(locales) && locales.length > 1">
+                            <div class="kiss-margin-small kiss-flex kiss-flex-middle kiss-visible-toggle" gap="xsmall" v-if="Array.isArray(locales) && locales.length > 1">
                                 <span class="kiss-badge kiss-badge-outline kiss-color-primary">{{ locale.i18n }}</span>
-                                <kiss-dropdown class="kiss-margin-small-left">
+                                <kiss-dropdown>
                                     <a class="kiss-invisible-hover kiss-color-muted" :ariaLabel="t('Copy value from another locale')" kiss-tooltip="right"><icon>content_copy</icon></a>
 
                                     <kiss-dropdownbox pos="left">
@@ -608,7 +608,7 @@ export default {
                                 </kiss-dropdown>
                             </div>
                             <div class="kiss-flex-1"></div>
-                            <div class="kiss-flex kiss-flex-middle app-fieldcontainer-visible-hover" gap="small" v-if="fieldActions.length">
+                            <div class="kiss-flex kiss-flex-middle app-fieldcontainer-visible-hover" gap="xsmall" v-if="fieldActions.length">
                                 <component :is="action.component" v-model="val[field.name+(locale.i18n === 'default' ? '': '_'+locale.i18n)]" :document="val" :field="field" :locale="locale.i18n" v-for="action in fieldActions"></component>
                             </div>
                         </div>