+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix MongoLite edge error (table exists)
 - Add more checks during Cockpit update
 - Make field actions available for multiple values fields
+- Locale data support for asset information
 
 ## 2.10.0 (2024-11-05)